// src/app/shared/models/flight.model.ts
export class Flight {
    idFlight: number | undefined;
    companyName: string | undefined;
    flightType: string | undefined;
    travelType: string | undefined;
    departureDate: Date | undefined;
    departureTime: Date | undefined;
    arrivaleDate: Date | undefined;
    arrivaleTime: Date | undefined;
    backDate: Date | undefined;
    backTime: Date | undefined;
    departureLocation: string | undefined;
    arrivalLocation: string | undefined;
    flightDuration: Date | undefined;
    connectionDuration: Date | undefined;
    aircraftType: string | undefined;
}